Check the whole line for wrap-introduced fake table delimiter rows
