+++ src/language-markdown/print/word.js
@@ -1,9 +1,6 @@
 import { PUNCTUATION_REGEXP } from "../constants.evaluate.js";
 import { isAutolink, isNewLine } from "../utilities.js";
-import {
-  isFakeTableDelimiterRow,
-  tableDelimiterRowStartRegex,
-} from "../utilities/is-fake-table-delimiter-row.js";
+import { isFakeTableDelimiterRowLine } from "../utilities/is-fake-table-delimiter-row.js";
 
 const fakeSetextHeaderRegex = /^(?:=+|-+)$/;
 
@@ -39,7 +36,7 @@ function printWord(path, options) {
       options.proseWrap === "preserve" &&
       path.parent.type === "sentence" &&
       isNewLine(path.previous) &&
-      isFakeTableDelimiterRowLine(path)
+      isFakeTableDelimiterRowLine(path.siblings, path.index)
     ) {
       // escape indented pseudo table delimiter row, e.g. `| x | y |↵␣␣␣␣| --- | --- |`
       return `\\${text}`;
@@ -83,38 +80,6 @@ function printWord(path, options) {
   );
 
   return text;
-}
-
-/**
- * A delimiter row with spaces around its cells, e.g. `| --- | --- |`, is
- * split into several sibling word nodes by the whitespace in between, so
- * checking the first word alone (`|`) misses it. Rebuild the line the word
- * starts and check that instead; escaping the leading word is still enough
- * to keep the whole line from parsing as a delimiter row.
- *
- * @param {AstPath} path
- * @returns {boolean}
- */
-function isFakeTableDelimiterRowLine(path) {
-  const { siblings, index } = path;
-  const words = [];
-  for (let i = index; i < siblings.length; i++) {
-    const sibling = siblings[i];
-    if (sibling.type === "whitespace") {
-      if (isNewLine(sibling)) {
-        break;
-      }
-      continue;
-    }
-    if (
-      sibling.type !== "word" ||
-      !tableDelimiterRowStartRegex.test(sibling.value)
-    ) {
-      break;
-    }
-    words.push(sibling.value);
-  }
-  return isFakeTableDelimiterRow(words.join(" "));
 }
 
 /**
