+++ src/language-markdown/utilities/is-fake-table-delimiter-row.js
@@ -1,3 +1,5 @@
+import { isNewLine } from "../utilities.js";
+
 const tableDelimiterCellRegex = /^ *:?-+:? *$/;
 const tableDelimiterRowStartRegex = /^[ |:-]+$/;
 
@@ -41,4 +43,40 @@ function splitCells(row) {
   return cells;
 }
 
-export { isFakeTableDelimiterRow, tableDelimiterRowStartRegex };
+/**
+ * A delimiter row with spaces around its cells, e.g. `| --- | --- |`, is
+ * split into several sibling word nodes by the whitespace in between, so
+ * checking a single word alone (`|`) misses it. Starting at `startIndex`,
+ * collect consecutive delimiter-row-shaped words up to the next newline and
+ * check the reconstructed line instead.
+ *
+ * @param {import("../utilities.js").TextNode[]} siblings
+ * @param {number} startIndex
+ * @returns {boolean}
+ */
+function isFakeTableDelimiterRowLine(siblings, startIndex) {
+  const words = [];
+  for (let i = startIndex; i < siblings.length; i++) {
+    const sibling = siblings[i];
+    if (sibling.type === "whitespace") {
+      if (isNewLine(sibling)) {
+        break;
+      }
+      continue;
+    }
+    if (
+      sibling.type !== "word" ||
+      !tableDelimiterRowStartRegex.test(sibling.value)
+    ) {
+      break;
+    }
+    words.push(sibling.value);
+  }
+  return isFakeTableDelimiterRow(words.join(" "));
+}
+
+export {
+  isFakeTableDelimiterRow,
+  isFakeTableDelimiterRowLine,
+  tableDelimiterRowStartRegex,
+};
