+++ tests/format/markdown/table/not-table-prose-wrap-always/not-table.md
@@ -1,1 +1,4 @@
 |x|y| |---|---|
+
+|a|b|
+    | --- | --- |
